backend/requirements.txt
Flask
flask-mysqldb
flask-cors
mysql-connector-python
flask_restful